requirements.txt
redis>=5.0.0
pymongo>=4.6.0
cassandra-driver>=3.28.0
neo4j>=5.15.0
requests>=2.31.0
